Refactor: move useSmartBack to common hooks folder
